refactor(profile): extract store to separate file and rename config files
Separates storage concerns from configuration model:
- Extract ProfileStore and methods to store.go
- Rename profile.go to config.go
- Rename profile_test.go to config_test.go
- Move ProfileStore tests to store_test.go

This is pure refactoring with no behavior change, preparing the
architecture for future pipeline profiles support.

Related: cb-s9n
